displayio: fix BMP parsing of CORE headers and 32-bit fields
A BITMAPCOREHEADER was accepted and then read entirely through Windows
INFOHEADER offsets. CORE stores 16-bit width and height at bytes 18 and 20
and the bit count at 24; the parser read a 32-bit width from 18, a height
from 22 and the depth from 28, then read a compression field at byte 30 --
which for a 12-byte header is stack the file never filled. Its palette is
3-byte RGBTRIPLEs, read as 4-byte RGBQUADs, which shifted every colour and
ran on into the pixel data. Every field of a CORE bitmap was wrong.

Width, height, the data offset and the stride all come from 32-bit header
fields but were stored in 16 bits, so a valid top-down BMP -- written with
a negative height -- became a huge positive one, and a file whose pixel
data starts past 65535 seeked elsewhere.

read_word promoted its high half to signed int before shifting, so any
value from 0x8000 up was undefined behaviour.

Also zeroes the header buffer, restricts the bit depth to the documented
set, and casts before the bounds comparison in get_pixel.

+++ shared-module/displayio/OnDiskBitmap.h
@@ -15,10 +15,15 @@
 
 typedef struct {
     mp_obj_base_t base;
-    uint16_t width;
-    uint16_t height;
-    uint16_t data_offset;
-    uint16_t stride;
+    // CIRCUITPY-CHANGE: width, height and the data offset come out of a 32 bit
+    // header and the stride is computed from the width, but all four were stored
+    // in 16 bits. A negative height, which is how a valid top-down BMP is written,
+    // became a large positive one, and a file whose pixel data starts past 65535
+    // seeked somewhere else entirely.
+    uint32_t width;
+    uint32_t height;
+    uint32_t data_offset;
+    uint32_t stride;
     uint32_t r_bitmask;
     uint32_t g_bitmask;
     uint32_t b_bitmask;
